Cleanup overclocking logic (#607)
* Cleanup overclocking logic

* Fix Multiblock maintenance and overclock interaction

* Gather various overclocking logics into a single class
