+++ src/main/java/gregtech/api/capability/impl/MultiblockRecipeLogic.java
@@ -10,6 +10,7 @@ import gregtech.api.metatileentity.multiblock.MultiblockWithDisplayBase;
 import gregtech.api.metatileentity.multiblock.RecipeMapMultiblockController;
 import gregtech.api.recipes.Recipe;
 import gregtech.api.recipes.RecipeMap;
+import gregtech.api.recipes.recipeproperties.RecipePropertyStorage;
 import gregtech.common.ConfigHolder;
 import net.minecraft.util.Tuple;
 import net.minecraftforge.items.IItemHandlerModifiable;
@@ -18,6 +19,8 @@ import javax.annotation.Nonnull;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+
+import static gregtech.api.recipes.logic.OverclockingLogic.standardOverclockingLogic;
 
 public class MultiblockRecipeLogic extends AbstractRecipeLogic {
 
@@ -251,17 +254,28 @@ public class MultiblockRecipeLogic extends AbstractRecipeLogic {
     }
 
     @Override
-    protected int[] runOverclockingLogic(@Nonnull Recipe recipe, boolean negativeEU, int maxOverclocks) {
+    protected int[] runOverclockingLogic(RecipePropertyStorage propertyStorage, int recipeEUt, long maxVoltage, int recipeDuration, int maxOverclocks) {
         // apply maintenance penalties
         Tuple<Integer, Double> maintenanceValues = getMaintenanceValues();
 
         int[] overclock = null;
         if (maintenanceValues.getSecond() != 1.0)
-            overclock = overclockRecipe(recipe.getRecipePropertyStorage(), recipe.getEUt(), negativeEU, getMaxVoltage(),
-                    (int) Math.round(recipe.getDuration() * maintenanceValues.getSecond()), maxOverclocks);
+
+            overclock = standardOverclockingLogic(Math.abs(recipeEUt),
+                    maxVoltage,
+                    (int) Math.round(recipeDuration * maintenanceValues.getSecond()),
+                    getOverclockingDurationDivisor(),
+                    getOverclockingVoltageMultiplier(),
+                    maxOverclocks
+            );
 
         if (overclock == null)
-            overclock = overclockRecipe(recipe.getRecipePropertyStorage(), recipe.getEUt(), negativeEU, getMaxVoltage(), recipe.getDuration(), maxOverclocks);
+            overclock = standardOverclockingLogic(Math.abs(recipeEUt),
+                    maxVoltage,
+                    recipeDuration,
+                    getOverclockingDurationDivisor(),
+                    getOverclockingVoltageMultiplier(),
+                    maxOverclocks);
 
         if (maintenanceValues.getFirst() > 0)
             overclock[1] = (int) (overclock[1] * (1 + 0.1 * maintenanceValues.getFirst()));
