+++ src/main/java/gregtech/api/capability/impl/PrimitiveRecipeLogic.java
@@ -4,8 +4,11 @@ import gregtech.api.GTValues;
 import gregtech.api.metatileentity.multiblock.RecipeMapPrimitiveMultiblockController;
 import gregtech.api.recipes.Recipe;
 import gregtech.api.recipes.RecipeMap;
+import gregtech.api.recipes.recipeproperties.RecipePropertyStorage;
 
 import javax.annotation.Nonnull;
+
+import static gregtech.api.recipes.logic.OverclockingLogic.standardOverclockingLogic;
 
 /**
  * Recipe Logic for a Multiblock that does not require power.
@@ -42,10 +45,10 @@ public class PrimitiveRecipeLogic extends AbstractRecipeLogic {
     }
 
     @Override
-    protected int[] runOverclockingLogic(@Nonnull Recipe recipe, boolean negativeEU, int maxOverclocks) {
+    protected int[] runOverclockingLogic(RecipePropertyStorage propertyStorage, int recipeEUt, long maxVoltage, int recipeDuration, int maxOverclocks) {
         return standardOverclockingLogic(1,
                 getMaxVoltage(),
-                recipe.getDuration(),
+                recipeDuration,
                 getOverclockingDurationDivisor(),
                 getOverclockingVoltageMultiplier(),
                 maxOverclocks
